// app/model/dto/EspecieDTO.ts
export interface EspecieDTO {
    id?: number | null;
    nombre: string | null;
    clasificacion: string | null;
    cargo: string | null;
    altura_prom: number | null;
    color_cabello: string | null;
    color_piel: string | null;
    color_ojos: string | null;
    promedio_vida: number | null;
    idioma: string | null;
}