// app/model/dto/VehiculoDTO.ts
export interface VehiculoDTO {
    id?: number;
    nombre?: string;
    altura?: number;
    modelo?: string;
    fabricante?: string;
    precio?: number;
    longitud?: number;
    velocidad_max?: number;
    tripulacion?: number;
    pasajeros?: number;
    capacidad_carga?: number;
    combustible?: string;
    tipo?: string;
    
}